requirements.txt
streamlit
pandas
reportlab
python-escpos
